(content): refactoring sections; extract Search Input to an outsider component
